updated app name and rollup config
