+++ src/popup/App.jsx
@@ -1,0 +1,3 @@
+export default function App() {
+  return <div>Hello World</div>;
+}
