--- src/popup/index.js
@@ -1,4 +1,0 @@
-import { createRoot } from 'react-dom/client';
-
-const root = createRoot(document.getElementById('root'));
-root.render(<h1>Hello, world</h1>);
