+++ src/popup/menu.html
@@ -13,11 +13,11 @@
     />
 
     <link rel="stylesheet" href="menu.css" />
-    <script type="module" src="menu.js"></script>
 
     <title>Border Patrol</title>
   </head>
   <body>
     <div id="root"></div>
+    <script src="menu.js"></script>
   </body>
 </html>
